+++ src_cpp/src/Engine/game/map/Density.cpp
@@ -29,30 +29,30 @@ int Density::maxLvl {3};
  * 根据 ecoObj 预设的方案 来配置 density.lvl
  */
 void Density::set( const IntVec2 &fieldMPos_, 
-                    const float &ecoObj_densitySeaLvlOff_,
-                    const std::vector<float> *ecoObj_densityDivideValsPtr_ ){
+                    const double &ecoObj_densitySeaLvlOff_,
+                    const std::vector<double> *ecoObj_densityDivideValsPtr_ ){
     //-----------//
     //   seaLvl
     //-----------//
-    float seaLvlOff = ecoObj_densitySeaLvlOff_; //- [-20.0, 20.0]
-    float freqSeaLvl = 0.003125f * 0.5f; //- 10*10 个 chunk，构成一个 perlin 晶格
-    float pnValSeaLvl = simplex_noise2( fieldMPos_.x * freqSeaLvl, 
-                                        fieldMPos_.y * freqSeaLvl ) * 15.0f + seaLvlOff; // [-20.0, 20.0]
+    double seaLvlOff = ecoObj_densitySeaLvlOff_; //- [-20.0, 20.0]
+    double freqSeaLvl = 0.003125 * 0.5; //- 10*10 个 chunk，构成一个 perlin 晶格
+    double pnValSeaLvl = simplex_noise2( fieldMPos_.x * freqSeaLvl, 
+                                        fieldMPos_.y * freqSeaLvl ) * 15.0 + seaLvlOff; // [-20.0, 20.0]
     //-----------//
     //   pnVal
     //-----------//
-    const glm::vec2 &densitySeed_pposOff = esrc::get_gameSeed().get_densitySeed_pposOff();
-    float freq = 0.03125f * 0.5f; //- 1*1 个 chunk，构成一个 perlin 晶格
-    float x = static_cast<float>(fieldMPos_.x) + densitySeed_pposOff.x;
-    float y = static_cast<float>(fieldMPos_.y) + densitySeed_pposOff.y;
-    y *= 1.3f; //- 在纵向上适当压缩
-    float pnVal = simplex_noise2(x*freq, y*freq) * 100.0f - pnValSeaLvl; // [-120.0, 120.0]
+    const glm::dvec2 &densitySeed_pposOff = esrc::get_gameSeed().get_densitySeed_pposOff();
+    double freq = 0.03125 * 0.5; //- 1*1 个 chunk，构成一个 perlin 晶格
+    double x = static_cast<double>(fieldMPos_.x) + densitySeed_pposOff.x;
+    double y = static_cast<double>(fieldMPos_.y) + densitySeed_pposOff.y;
+    y *= 1.3; //- 在纵向上适当压缩
+    double pnVal = simplex_noise2(x*freq, y*freq) * 100.0 - pnValSeaLvl; // [-120.0, 120.0]
 
-    if( pnVal > 100.0f ){
-        pnVal = 100.0f;
+    if( pnVal > 100.0 ){
+        pnVal = 100.0;
     }
-    if( pnVal < -100.0f ){
-        pnVal = 100.0f;
+    if( pnVal < -100.0 ){
+        pnVal = 100.0;
     } // [-100.0, 100.0]
 
     //-----------//
